implemented picture upload, security checks & refactored code

+++ app/controllers/Product.php
@@ -10,24 +10,28 @@
             }
 
             #[\app\filters\LoginAsStore]
-            public function create($product_id) {
-                if (!isset($_POST['action'])) { 
-                    $this->view('Product/create');
+            public function create($store_id) {
+                if ($store_id == $_SESSION['store_id']) {
+                    if (!isset($_POST['action'])) { 
+                        $this->view('Product/create');
+                    }
+                    else {  //implement pictures later
+                        $filename = Main::imageUpload("product_image");
+                        $product = new \app\models\Product();
+                        $product->store_id = $store_id;
+                        $product->product_name = $_POST['product_name'];
+                        $product->product_image = $filename; 
+                        $product->product_availability = $this->getAvailability($_POST['product_quantity']);
+                        $product->product_quantity = $_POST['product_quantity'];
+                        $product->product_price = $_POST['product_price'];
+                        $product->product_description = $_POST['product_description'];
+
+                        $product->insert();
+                        header("location:/Store/index/". $_SESSION['store_id']);
+                    }
                 }
-                else {  //implement pictures later + fix picture folder structure (per controller)
-			        $filename = 'blank.png';
-                    $product = new \app\models\Product();
-                    $product->product_id = $product_id;
-                    $product->store_id = $_SESSION['store_id'];
-                    $product->product_name = $_POST['product_name'];
-                    $product->product_image = $filename; //change this later
-                    $product->product_availability = $this->getAvailability($_POST['product_quantity']);
-                    $product->product_quantity = $_POST['product_quantity'];
-                    $product->product_price = $_POST['product_price'];
-                    $product->product_description = $_POST['product_description'];
-                    $product->insert();
+                else 
                     header("location:/Store/index/". $_SESSION['store_id']);
-                }
             }
 
             private function getAvailability($quantity) {
@@ -41,30 +45,45 @@
             public function update($product_id) {
                 $product = new \app\models\Product(); 
                 $product = $product->get($product_id);
-                if (!isset($_POST['action'])) {	
-                    $this->view('product/update', $product);
+                if ($product->store_id == $_SESSION['store_id']) {
+                    if (!isset($_POST['action'])) {	
+                        $this->view('Product/update', $product);
+                    }
+                    else {
+                        $filename = Main::imageUpload("product_image");
+                        if (!$filename) {
+                            if ($product->product_image != 'blank.jpg')
+                                unlink('app\\pictures\\' . $product->product_image);
+
+                            $product->product_imagee = $filename;
+                        }
+
+                        $product->product_name = $_POST['product_name'];
+                        $product->product_quantity= $_POST['product_quantity'];
+                        $product->product_price = $_POST['product_price'];
+                        $product->product_description = $_POST['product_description'];
+
+                        $product->update($product_id);        
+                        header('location:/Store/index/' . $_SESSION['store_id']);
+                    }
                 }
-                else {
-                    $product->product_name = $_POST['product_name'];
-                    $product->product_quantity= $_POST['product_quantity'];
-                    $product->product_price = $_POST['product_price'];
-                    // $product->product_list = ; //redo this later, product/create link, add to array, then concatenate to string using , as delimiter for ids
-                    $product->product_description = $_POST['product_description'];
-                    $product->update($product_id);
-                    // $product_id = $newproduct->getLast()->product_id; //test this
-                    // $_SESSION['product_id'] = $product_id;
-                   
-                    header('location:/store/index/' . $product->store_id);
-                }
+                else
+                    header('location:/Store/index/' . $_SESSION['store_id']);
             }
 
             #[\app\filters\LoginAsStore]
             public function delete($product_id) {
                 $product = new \app\models\Product(); 
                 $product = $product->get($product_id);
-              
-                $product->delete(); 
-                header('location:/store/index/' . $product->store_id);
+
+                if ($product->store_id == $_SESSION['session_id']) { //checking if product belongs to store
+                    if ($product->product_image != 'blank.jpg')
+                        unlink('app\\pictures\\' . $product->product_image);
+                        
+                    $product->delete();
+                }
+
+                header('location:/Store/index/' . $_SESSION['store_id']);
             }
             
         }
