+++ app/controllers/Store.php
@@ -11,57 +11,68 @@
             
             #[\app\filters\Login]
             public function create($user_id) {
-                if (!isset($_POST['action'])) {	//display the view if I don't submit the form
-                    $this->view('Store/create');
-                }
-                else {	//process the data when the form has been submitted, id, title, text
-                    $newStore = new \app\models\Store();
-                    $newStore->store_name = $_POST['store_name'];
-                    $newStore->store_address = $_POST['store_address'];
-                    //$newStore->product_list = ; //redo this later, product/create link, add to array, then concatenate to string using , as delimiter for ids
-                    $newStore->description = $_POST['description'];
+                if ($user_id == $_SESSION['user_id']) {
+                    if (!isset($_POST['action'])) {	//display the view if I don't submit the form
+                        $this->view('Store/create');
+                    }
+                    else {	//process the data when the form has been submitted, id, title, text
+                        $newStore = new \app\models\Store();
+                        $newStore->store_name = $_POST['store_name'];
+                        $newStore->store_address = $_POST['store_address'];
+                        $newStore->description = $_POST['description'];
 
-                    $newStore->insert($user_id);
-                    
+                        $newStore->insert($user_id);
 
-                    $store_id = $newStore->getLast()->store_id; //test this
-                    $_SESSION['store_id'] = $store_id;
-                    header('location:/Store/index/' . $store_id);
-                    //todo: make create view 
+                        $store_id = $newStore->getLast()->store_id; 
+                        $_SESSION['store_id'] = $store_id;
+                        header('location:/Store/index/' . $_SESSION['store_id']);
+                    }
                 }
+                else
+                    header('location:/User/index/' . $_SESSION['user_id']);
             }
 
             #[\app\filters\LoginAsStore]
             public function update($store_id) {
                 $store = new \app\models\Store(); 
-                $store = $store->get($store_id);
-                if (!isset($_POST['action'])) {	
-                    $this->view('Store/update', $store);
+                if ($store_id == $_SESSION['store_id']) {
+                    $store = $store->get($store_id);
+                    if (!isset($_POST['action'])) {	
+                        $this->view('Store/update', $store);
+                    }
+                    else {
+                        $store->store_name = $_POST['store_name'];
+                        $store->store_address = $_POST['store_address'];
+                        $store->description = $_POST['description'];
+
+                        $store->update($store_id);
+                        header('location:/Store/index/' . $_SESSION['store_id']);
+                    }
                 }
-                else {
-                    $store->store_name = $_POST['store_name'];
-                    $store->store_address = $_POST['store_address'];
-                    // $store->product_list = ; //redo this later, product/create link, add to array, then concatenate to string using , as delimiter for ids
-                    $store->description = $_POST['description'];
-                    $store->update($store_id);
-                    header('location:/Store/index/' . $store_id);
-                }
+                else 
+                    header('location:/Store/index/' . $_SESSION['store_id']);
             }
 
             #[\app\filters\LoginAsStore]
             public function delete($store_id) {
+                $user = new \app\models\User();
+                $user = $user->getById($_SESSION['user_id']);
                 $store = new \app\models\Store(); 
-                $store = $store->get($store_id);
+                if ($store_id == $_SESSION['store_id']) {
+                    $store = $store->get($store_id);
 
-                $products = new \app\models\Product(); 
-                $products = $products->getAllFromStore($store_id); 
+                    $products = new \app\models\Product(); 
+                    $products = $products->getAllFromStore($store_id); 
 
-                foreach ($products as $product){
-                    $product->delete(); 
+                    foreach ($products as $product) {
+                        $product->delete(); 
+                    }
+
+                    $user->updateUserType();
+                    $store->delete($store_id);
+                    header('location:/User/index/' . $_SESSION['user_id']);
                 }
-
-                $user::updateUserType();
-                $store->delete($store_id);
-                header('location:/User/index/' . $_SESSION['user_id']);
+                else 
+                    header('location:/Store/index/' . $_SESSION['store_id']);
             }
         }
